requirements.txt
certifi==2020.12.5
click==7.1.2
cloudinary==1.25.0
Flask-Cors==3.0.10
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.16.0
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
